PerformanceAssessment_App/pysimmusic-extended_rhythm_error_visualization/requirements.txt
numpy==1.16.4
Pygments==2.2.0
appdirs==1.4.3
decorator==4.0.11
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
py==1.4.33
pyparsing==2.2.0
PySoundFile==0.9.0.post1
pytest==3.0.7
scipy==1.2.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
Cython==0.25.2
click==6.7
sphinx==1.6.4
sphinx_rtd_theme==0.2.4
pymongo==3.5.1
boto3==1.4.7
pandas==0.21.0
scikit-learn==0.21.2
numpydoc==0.7.0
resampy==0.2.1
nwalign3==0.1.0
vamp==1.1.0
matplotlib==2.2.2
joblib==0.11
xmltodict==0.11.0
music21==5.1.0
Pillow==6.0.0
